refactor(app): share GameScreen shell hosts and densify near-cap suites
Extract game_screen_test_support for duplicated shell overrides, move
production panel fixtures under support/, and split narrow/production
mega-suites under the size gate (Refs #4013).
