+++ app/test/game_map_players_bar_narrow_test.dart
@@ -24,30 +24,20 @@
 
 import 'package:colonizethis_app/app.dart';
 import 'package:colonizethis_app/config/constants.dart';
-import 'package:colonizethis_app/config/themes.dart';
-import 'package:colonizethis_app/core/services/app_event_handler/app_event_handler_scope.dart';
-import 'package:colonizethis_app/core/services/game_service/game_service.dart';
 import 'package:colonizethis_app/features/game/screens/game/game_screen.dart'
-    show GameScreen, kGameMapPlayersBarKey;
+    show kGameMapPlayersBarKey;
 import 'package:colonizethis_app/features/game/widgets/shell/game_map_players_bar.dart';
 import 'package:colonizethis_app/features/game/widgets/shell/player_turn_event_feed.dart';
-import 'package:colonizethis_app/providers/app_event_bus_provider.dart';
-import 'package:colonizethis_app/providers/game_service_provider.dart';
-import 'package:colonizethis_app/providers/games_box_provider.dart';
-import 'package:colonizethis_app/providers/games_provider.dart';
-import 'package:colonizethis_app/providers/home_fleet_cargo_provider.dart';
-import 'package:colonizethis_app/providers/map_view_provider.dart';
-import 'package:colonizethis_app/providers/treasury_summary_provider.dart';
-import 'package:colonizethis_map/colonizethis_map.dart' show InitGameMapViewData;
+import 'package:colonizethis_map/colonizethis_map.dart'
+    show InitGameMapViewData;
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_save/colonizethis_save.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
 import 'support/map_view_test_fixtures.dart';
+import 'support/game_screen_test_support.dart';
 import 'support/panel_test_fixtures.dart';
 
 void main() {
@@ -66,51 +56,18 @@ void main() {
     gamesBox = await Hive.openBox<dynamic>(HiveBoxNames.games);
   });
 
-  gameShellOverrides({required Game game}) => [
-      gamesBoxProvider.overrideWith((ref) => gamesBox),
-      gameServiceProvider.overrideWith(
-        (ref) => GameService(gamesBox, GameSaveAdapter()),
-      ),
-      currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-      currentOrdersProvider.overrideWith(
-        () => CurrentOrdersNotifier(const Orders()),
-      ),
-      mapViewDataProvider.overrideWith((ref) => mapViewData),
-      gameIdsWithIntroShownProvider.overrideWith(
-        () => GameIdsWithIntroShownNotifier({game.id}),
-      ),
-      appEventBusProvider.overrideWith((ref) {
-        final bus = AppEventBus.create();
-        ref.onDispose(bus.dispose);
-        return bus;
-      }),
-      homeFleetCargoSummaryProvider.overrideWith(
-        (ref) => const HomeFleetCargoSummary(used: 0, capacity: 0),
-      ),
-      treasurySummaryProvider.overrideWith(
-        (ref) => const TreasurySummary(treasury: 12345),
-      ),
-    ];
-
   Widget buildGameScreen({
     required Game game,
     required double width,
     required double height,
-  }) {
-    return ProviderScope(
-      overrides: gameShellOverrides(game: game),
-      child: AppEventHandlerScope(
-        child: MaterialApp(
-          navigatorKey: appNavigatorKey,
-          theme: AppThemes.colonial,
-          home: MediaQuery(
-            data: MediaQueryData(size: Size(width, height)),
-            child: const GameScreen(),
-          ),
-        ),
-      ),
-    );
-  }
+  }) => buildGameScreenHost(
+    gamesBox: gamesBox,
+    game: game,
+    mapViewData: mapViewData,
+    width: width,
+    height: height,
+    navigatorKey: appNavigatorKey,
+  );
 
   group('GameMapPlayersBar narrow-viewport gating (Refs #3898)', () {
     testWidgets(
@@ -127,11 +84,7 @@ void main() {
         addTearDown(tester.view.reset);
 
         await tester.pumpWidget(
-          buildGameScreen(
-            game: game,
-            width: narrowWidth,
-            height: 700,
-          ),
+          buildGameScreen(game: game, width: narrowWidth, height: 700),
         );
         await tester.pump();
 
@@ -156,11 +109,7 @@ void main() {
         addTearDown(tester.view.reset);
 
         await tester.pumpWidget(
-          buildGameScreen(
-            game: game,
-            width: narrowWidth,
-            height: 700,
-          ),
+          buildGameScreen(game: game, width: narrowWidth, height: 700),
         );
 
         const step = Duration(milliseconds: 50);
@@ -204,11 +153,7 @@ void main() {
         addTearDown(tester.view.reset);
 
         await tester.pumpWidget(
-          buildGameScreen(
-            game: game,
-            width: narrowWidth,
-            height: 700,
-          ),
+          buildGameScreen(game: game, width: narrowWidth, height: 700),
         );
         await tester.pump();
 
@@ -230,11 +175,7 @@ void main() {
         addTearDown(tester.view.reset);
 
         await tester.pumpWidget(
-          buildGameScreen(
-            game: baseGame,
-            width: wideWidth,
-            height: 700,
-          ),
+          buildGameScreen(game: baseGame, width: wideWidth, height: 700),
         );
         await tester.pump();
 
@@ -265,8 +206,7 @@ void main() {
           victory: VictoryState(
             winnerPlayerId: baseGame.players.first.id,
             type: VictoryType.military,
-            turnNumber:
-                baseGame.worldState.turnState.turnNumber,
+            turnNumber: baseGame.worldState.turnState.turnNumber,
           ),
         );
 
@@ -275,11 +215,7 @@ void main() {
         addTearDown(tester.view.reset);
 
         await tester.pumpWidget(
-          buildGameScreen(
-            game: victoryGame,
-            width: wideWidth,
-            height: 700,
-          ),
+          buildGameScreen(game: victoryGame, width: wideWidth, height: 700),
         );
         await tester.pump();
