+++ app/test/production_panel_available_alignment_golden_test.dart
@@ -18,7 +18,7 @@ import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'production_panel_test_fixtures.dart';
+import 'support/production_panel_test_support.dart';
 import 'support/golden_capture_harness.dart';
 import 'widget_test_pumps.dart';
