+++ app/test/production_panel_available_grid_test.dart
@@ -10,7 +10,7 @@ import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'production_panel_test_fixtures.dart';
+import 'support/production_panel_test_support.dart';
 import 'widget_test_pumps.dart';
 
 class _ProductionPanelGridTestWrapper extends StatefulWidget {
