+++ app/test/production_panel_normalized_recipe_labels_golden_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app/features/game/widgets/production/production_panel.dart';
 
-import 'production_panel_test_fixtures.dart';
+import 'support/production_panel_test_support.dart';
 import 'support/golden_capture_harness.dart';
 import 'widget_test_pumps.dart';
