+++ app/test/production_screen_integration_test.dart
@@ -15,7 +15,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'widget_test_pumps.dart';
-import 'production_panel_test_fixtures.dart';
+import 'support/production_panel_test_support.dart';
 
 class _SeededProductionDesiredOutputNotifier
     extends ProductionDesiredOutputNotifier {
