+++ app/test/production_screen_top_bar_test.dart
@@ -17,7 +17,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'support/app_shell_harness.dart';
-import 'production_panel_test_fixtures.dart';
+import 'support/production_panel_test_support.dart';
 import 'widget_test_pumps.dart';
 
 void main() {
